% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_frequency.r
\name{T_frequency}
\alias{T_frequency}
\title{A function characterizes frequency of stream thermal regime}
\usage{
T_frequency(sitedata, cT = 16, TlengthPortion = 2/3, SeasonSp = c(3, 4,
  5), SeasonSu = c(6, 7, 8), SeasonFa = c(9, 10, 11), SeasonWi = c(12, 1,
  2))
}
\arguments{
\item{sitedata}{stream monitoring site data in SiteID, Date(in as.Date format), 
MaxT, MinT, MeanT.}

\item{cT}{critical temperature of interest, default = 16 (Celsius)}

\item{TlengthPortion}{portion of length that required for calculating metric summaries,
defaulty=2/3}

\item{SeasonSp}{define spring season, default as March, April, and May, c(3,4,5)}

\item{SeasonSu}{define summer season, default as June, July, August, c(6,7,8)}

\item{SeasonFa}{define fall season, default as September, October, November, c(9,10,11)}

\item{SeasonWi}{define winter season, default as December, Januray, February, c(12,1,2)}
}
\description{
This function summarize metrics of stream thermal frequency.
}
\examples{
install.packages("dataRetrieval")
library(dataRetrieval)
ExUSGSStreamTemp<-readNWISdv("01382310","00010","2011-01-01","2011-12-31",c("00001","00002","00003"))
sitedata<-subset(ExUSGSStreamTemp, select=c("site_no","Date","X_00010_00001","X_00010_00002","X_00010_00003"))
names(sitedata)<-c("siteID","Date","MaxT","MinT","MeanT")
T_frequency(sitedata)
}
\keyword{timing}

